requirements.txt
langchain>=0.3.14
langchain-openai>=0.2.15
openai>=1.58.1
python-dotenv>=1.0.1
fastapi>=0.115.6
uvicorn>=0.34.0
pydantic>=2.10.4
gunicorn>=21.2.0
gevent>=24.11.1
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
botbuilder-integration-aiohttp>=4.15.0
aiohttp>=3.9.0